Added fluent simulation time, added double for more precision

+++ module.cpp
@@ -1,6 +1,6 @@
 #include "module.hpp"
 
-Module::Module(std::string name, float hull) : 
+Module::Module(std::string name, double hull) : 
 enabled(true), 
 name(name), 
 hull(hull),
